services/requirements.txt
langchain
fastapi
uvicorn
numpy 
pydantic
python-multipart
langchain_groq
python-dotenv
groq
pymongo
motor